Removing Impl sufix, moving folders as service should be another layer but not inside domain, and all entities are directly inside domain.
