+++ LibraryProjectt/src/main/java/com/library/domain/repositories/BookLeadingRepository.java
@@ -1,6 +1,6 @@
 package com.library.domain.repositories;
 
-import com.library.domain.entities.BookLeading;
+import com.library.domain.BookLeading;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
