+++ LibraryProjectt/src/main/java/com/library/domain/repositories/BookRepository.java
@@ -1,7 +1,7 @@
 package com.library.domain.repositories;
 
 
-import com.library.domain.entities.Book;
+import com.library.domain.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
