+++ LibraryProjectt/src/main/java/com/library/domain/repositories/ReaderRepository.java
@@ -1,7 +1,7 @@
 package com.library.domain.repositories;
 
 
-import com.library.domain.entities.Reader;
+import com.library.domain.Reader;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
