--- LibraryProjectt/src/main/java/com/library/domain/services/BookLeadingService.java
@@ -1,15 +1,0 @@
-package com.library.domain.services;
-
-
-import com.library.domain.entities.BookLeading;
-
-import java.util.List;
-
-public interface BookLeadingService {
-
-    BookLeading addBookLeading(BookLeading bookLeading);
-
-    List<BookLeading> getAll();
-
-    void delete(long id);
-}
